requirements.txt
python-decouple
pyrogram
requests
tgcrypto